Bring back geometry hints
Having size hints is great for aesthetics of floating windows.

In 4ba98271ff8389141dc084e3915e4faa66fb21f9, we removed calls to
vte_terminal_set_geometry_hints_for_window(). Our commit references
the following commit in VTE:

> commit e2108c7db3c1136777212ed33ba8ab716e87b4f2
> Author: Christian Persch <chpe@src.gnome.org>
> Date:   Sat Oct 21 22:30:29 2017 +0200
>
>     widget: Deprecate geometry helpers
>
>     They can't work correctly since gtk+ removed geometry support,
>     so all apps need to handle this themselves.

The release notes of GTK 3.20 said:

> Geometry handling in GtkWindow has been removed. If you are using the
> functions gtk_window_resize_to_geometry(),
> gtk_window_set_default_geometry(), gtk_window_parse_geometry() or
> gtk_window_set_geometry_hints(), you may need to make some changes to
> your code.

gtk_window_set_geometry_hints() is the function that VTE is using
internally. However, the function itself has never been removed from
GTK3 as far as I can tell and it is not deprecated in GTK3, either. It
is no longer present in GTK4, though.

It's certainly deprecated in VTE, i.e. if you use
vte_terminal_set_geometry_hints_for_window() you will get a warning. I
do not want to use this function, maybe it vanishes in a future release.

This is all a bit confusing and I think the term "geometry" has special
meaning in the world of GTK. I suspect that it's short for "geometry
widget" and GTK3 only removed support for that particular feature, not
size hints in general. In the old GTK2 docs (on some obscure website,
not on the official docs), "geometry widget" is explained like this:

> The canonical use-case for gtk_window_set_geometry_hints() is to get a
> terminal widget to resize properly. Here, the terminal text area should
> be the geometry widget; GtkWindow will then automatically set the base
> size to the size of other widgets in the terminal window, such as the
> menubar and scrollbar.

And the GTK3 docs now explain the parameter `geometry_widget` of
gtk_window_set_geometry_hints() like so:

> Widget the geometry hints used to be applied to or NULL. Since 3.20
> this argument is ignored and GTK behaves as if NULL was set.

So I guess this whole thing doesn't matter for us at all, because we
have no other widgets.

I don't remember anymore if this was actually broken in 2018 or if we
just removed the calls due to the deprecation warnings. Let's just bring
it back and try to build a working solution.

(Surprisingly, this even works on Wayland. No idea why, though, and it's
very, very sluggish on Sway. labwc is fine.)